More caution with 'rm -rf'

+++ tests/ci/common_posix_setup.sh
@@ -57,7 +57,7 @@ fi
 
 function run_build {
   local cflags=("$@")
-  rm -rf "$BUILD_ROOT"
+  rm -rf "${BUILD_ROOT:?}"
   mkdir -p "$BUILD_ROOT"
   cd "$BUILD_ROOT" || exit 1
 
@@ -207,7 +207,7 @@ function aws_lc_build() {
   ${CMAKE_COMMAND} ${AWS_LC_DIR} -GNinja "-B${BUILD_FOLDER}" "-DCMAKE_INSTALL_PREFIX=${INSTALL_FOLDER}" "${@:4}"
   ninja -C ${BUILD_FOLDER} install
   ls -R ${INSTALL_FOLDER}
-  rm -rf ${BUILD_FOLDER:?}/*
+  rm -rf "${BUILD_FOLDER:?}"/*
 }
 
 function print_executable_information {
