+++ tests/ci/docker_images/dependencies/build_cryptofuzz_modules.sh
@@ -61,7 +61,7 @@ env CRYPTOFUZZ_SRC "$CRYPTOFUZZ_SRC"
 
 # Cryptofuzz builds its modules into $CRYPTOFUZZ_SRC/modules that includes everything it needs, deleting the module source
 # code saves a substantial amount of space in the docker image
-rm -rf "$MODULES_ROOT"
+rm -rf "${MODULES_ROOT:?}"
 
 # Prebuild the required libcpu_features to save time
 cd "$CRYPTOFUZZ_SRC"
